stratify-backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2
tensorflow-cpu==2.15.0
joblib==1.3.2
python-telegram-bot==20.7
yfinance==0.2.32
google-generativeai==0.3.2
requests==2.31.0
